requirements.txt
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
Flask==0.11.1
Flask-Cors==2.1.2
flasgger==0.9.2
Flask-Limiter==1.0.1
ccxt==1.17.434
qrcode==6.1
pycryptodome==3.8.0
base58==0.2.5
sendypy==0.1.5b0
psycopg2-binary==2.8.2
PALs==0.2.0
peewee==3.9.4
dsnparse==0.1.12
